requirements.txt
discord.py==1.5.1
rockset==0.7.49
requests==2.25.0